Move dashboard stories into a new directory since the test data is getting large
